+++ NEWS.md
@@ -1,6 +1,6 @@
 # git2rdata 0.4.1
 
-* Add `update_description()` to update the description of a `git2rdata` object.
+* Add `update_metadata()` to update the description of a `git2rdata` object.
   See `vignette("metadata")` for more details.
 * Update the checklist and pkgdown infrastructure.